refactor: inject data provider to avoid globals
