+++ lib/yargs/commands/file.js
@@ -1,21 +1,20 @@
 'use strict'
 
 const path = require('path')
-const traceLocationProvider = require('../middlewares/trace-locations-provider')
 const { createEventQRCode } = require('../../index')
 
 const supportedExtensions = ['png', 'svg']
 
-module.exports = () => {
+module.exports = ({ dataProvider }) => {
   return {
     command: ['file'],
     desc: 'write QR code to a file',
     builder: yargs => {
-      traceLocationProvider.modifyBuilder(yargs)
+      dataProvider.builder(yargs)
     },
     handler: async argv => {
-      const { traceLocations } = argv
-      const allTraceLocationsCreated = traceLocations
+      const data = dataProvider.getData(argv)
+      const allQRCodesCreated = data
         .map(async ({ locationData, vendorData, filepath }) => {
           const eventQRCode = createEventQRCode({ locationData, vendorData })
           const ext = path.extname(filepath).substr(1)
@@ -28,9 +27,9 @@ module.exports = () => {
 
           console.log(`Created ${filepath}`)
         })
-      await Promise.all(allTraceLocationsCreated)
+      await Promise.all(allQRCodesCreated)
 
-      console.log(`Done creating ${traceLocations.length} QR code(s).`)
+      console.log(`Done creating ${data.length} QR code(s).`)
     }
   }
 }
