+++ lib/yargs/commands/poster.js
@@ -1,19 +1,18 @@
 'use strict'
 
 const path = require('path')
-const traceLocationProvider = require('../middlewares/trace-locations-provider')
 const { createEventQRCode } = require('../../index')
 
-module.exports = () => {
+module.exports = ({ dataProvider }) => {
   return {
     command: ['poster'],
     desc: 'create poster with QR code',
     builder: yargs => {
-      traceLocationProvider.modifyBuilder(yargs)
+      dataProvider.builder(yargs)
     },
     handler: async argv => {
-      const { traceLocations } = argv
-      const allTraceLocationsCreated = traceLocations
+      const data = dataProvider.getData(argv)
+      const allPostersCreated = data
         .map(async ({ locationData, vendorData, filepath }) => {
           const eventQRCode = createEventQRCode({ locationData, vendorData })
 
@@ -22,9 +21,9 @@ module.exports = () => {
 
           console.log(`Created ${filepath}`)
         })
-      await Promise.all(allTraceLocationsCreated)
+      await Promise.all(allPostersCreated)
 
-      console.log(`Done creating ${traceLocations.length} QR code poster(s).`)
+      console.log(`Done creating ${data.length} QR code poster(s).`)
     }
   }
 }
