+++ lib/yargs/commands/terminal.js
@@ -1,26 +1,25 @@
 'use strict'
 
-const traceLocationProvider = require('../middlewares/trace-locations-provider')
 const { createEventQRCode } = require('../../index')
 
-module.exports = () => {
+module.exports = ({ dataProvider }) => {
   return {
     command: ['terminal'],
     desc: 'print a QR code on the terminal',
     builder: yargs => {
-      traceLocationProvider.modifyBuilder(yargs)
+      dataProvider.builder(yargs)
     },
     handler: async argv => {
-      const { traceLocations } = argv
-      const allTraceLocationsCreated = traceLocations
+      const data = dataProvider.getData(argv)
+      const allQRCodesCreated = data
         .map(async ({ locationData, vendorData }) => {
           const eventQRCode = createEventQRCode({ locationData, vendorData })
 
           console.log(await eventQRCode.toString())
         })
-      await Promise.all(allTraceLocationsCreated)
+      await Promise.all(allQRCodesCreated)
 
-      console.log(`Done creating ${traceLocations.length} QR code(s).`)
+      console.log(`Done creating ${data.length} QR code(s).`)
     }
   }
 }
